refactor: organizando arquivos do front em pastas
